+++ src/GuildsApp.tsx
@@ -5,7 +5,6 @@ import {
   useHistory,
   Redirect,
 } from 'react-router-dom';
-import { EtherSWRConfig } from 'ether-swr';
 import { ThemeProvider } from 'styled-components';
 
 import { Container } from './components/Guilds/common/Layout';
@@ -18,21 +17,17 @@ import theme from './theme/light.json';
 import { GuildsContextProvider, TransactionsProvider } from 'contexts/Guilds';
 import WalletWeb3Manager from './components/Guilds/Web3Manager/WalletWeb3Manager';
 import GlobalErrorBoundary from './components/Guilds/ErrorBoundary/GlobalErrorBoundary';
-import useJsonRpcProvider from 'hooks/Guilds/web3/useJsonRpcProvider';
-import ERC20GuildContract from 'contracts/ERC20Guild.json';
 
 import ProposalTypes from 'components/Guilds/ProposalTypes';
 import { ProposalTypesConfig } from 'configs/proposalTypes';
 
 import ToastNotificationContainer from './components/Guilds/ToastNotifications/ToastNotificationContainer';
-import loggerMiddleware from './hooks/Guilds/ether-swr/middleware/logger';
+import EtherSWRManager from 'components/Guilds/EtherSWRManager';
 
 const GuildsApp = () => {
   const history = useHistory();
 
   const isTestingEnv = !window.location?.hostname?.startsWith('dxvote.eth');
-  const provider = useJsonRpcProvider();
-
   if (!isTestingEnv) {
     history.push('/');
     return null;
@@ -42,84 +37,42 @@ const GuildsApp = () => {
     <ThemeProvider theme={theme}>
       <HashRouter basename="/guilds">
         <GlobalErrorBoundary>
-          <WalletWeb3Manager>
-            <TransactionsProvider>
-              <GuildsContextProvider>
-                <GlobalStyle />
-                <Header />
-                <Container>
-                  <Switch>
-                    <Redirect
-                      exact
-                      from="/"
-                      to="/rinkeby/0x9cdc16b5f95229b856cba5f38095fd8e00f8edef"
-                    />
-                    <Redirect
-                      exact
-                      from="/:chain_name"
-                      to="/:chain_name/0x9cdc16b5f95229b856cba5f38095fd8e00f8edef"
-                    />
-                    <Route exact path="/:chain_name/:guild_id">
-                      <EtherSWRConfig
-                        value={{
-                          web3Provider: provider,
-                          ABIs: new Map([
-                            [
-                              // we can move this probably to a hook to reduce repeat ourselves in each route.
-                              '0x9cdc16b5f95229b856cba5f38095fd8e00f8edef',
-                              ERC20GuildContract.abi,
-                            ],
-                          ]),
-                          refreshInterval: 30000,
-                          use: [loggerMiddleware],
-                        }}
-                      >
+          <EtherSWRManager>
+            <WalletWeb3Manager>
+              <TransactionsProvider>
+                <GuildsContextProvider>
+                  <GlobalStyle />
+                  <Header />
+                  <Container>
+                    <Switch>
+                      <Redirect
+                        exact
+                        from="/"
+                        to="/rinkeby/0x9cdc16b5f95229b856cba5f38095fd8e00f8edef"
+                      />
+                      <Redirect
+                        exact
+                        from="/:chain_name"
+                        to="/:chain_name/0x9cdc16b5f95229b856cba5f38095fd8e00f8edef"
+                      />
+                      <Route exact path="/:chain_name/:guild_id">
                         <GuildsPage />
-                      </EtherSWRConfig>
-                    </Route>
-                    <Route exact path="/:chain_name/:guild_id/proposalType">
-                      <EtherSWRConfig
-                        value={{
-                          web3Provider: provider,
-                          ABIs: new Map([
-                            [
-                              // we can move this probably to a hook to reduce repeat ourselves in each route.
-                              '0x9cdc16b5f95229b856cba5f38095fd8e00f8edef',
-                              ERC20GuildContract.abi,
-                            ],
-                          ]),
-                          refreshInterval: 30000,
-                        }}
-                      >
+                      </Route>
+                      <Route exact path="/:chain_name/:guild_id/proposalType">
                         <ProposalTypes data={ProposalTypesConfig} />
-                      </EtherSWRConfig>
-                    </Route>
-                    <Route path="/:chain_name/:guild_id/proposal/:proposal_id">
-                      <EtherSWRConfig
-                        value={{
-                          web3Provider: provider,
-                          ABIs: new Map([
-                            [
-                              '0x9cdc16b5f95229b856cba5f38095fd8e00f8edef',
-                              ERC20GuildContract.abi,
-                            ],
-                          ]),
-                          refreshInterval: 0,
-                          use: [loggerMiddleware],
-                        }}
-                      >
-                        {' '}
+                      </Route>
+                      <Route path="/:chain_name/:guild_id/proposal/:proposal_id">
                         <ProposalPage />
-                      </EtherSWRConfig>
-                    </Route>
-                    <Route path="/:chain_name/:guild_id/create/:proposal_type">
-                      <CreateProposalPage />
-                    </Route>
-                  </Switch>
-                </Container>
-              </GuildsContextProvider>
-            </TransactionsProvider>
-          </WalletWeb3Manager>
+                      </Route>
+                      <Route path="/:chain_name/:guild_id/create/:proposal_type">
+                        <CreateProposalPage />
+                      </Route>
+                    </Switch>
+                  </Container>
+                </GuildsContextProvider>
+              </TransactionsProvider>
+            </WalletWeb3Manager>
+          </EtherSWRManager>
         </GlobalErrorBoundary>
       </HashRouter>
