+++ src/hooks/Guilds/ether-swr/useGuildProposals.ts
@@ -1,4 +1,7 @@
 import useEtherSWR from './useEtherSWR';
+import ERC20GuildContract from 'contracts/ERC20Guild.json';
 
 export const useGuildProposals = (guildId: string) =>
-  useEtherSWR([guildId, 'getProposalsIds']);
+  useEtherSWR([guildId, 'getProposalsIds'], {
+    ABIs: new Map([[guildId, ERC20GuildContract.abi]]),
+  });
