+++ src/hooks/Guilds/ether-swr/useVotingPowerOf.ts
@@ -1,6 +1,7 @@
 import { BigNumber } from 'ethers';
 import { SWRResponse } from 'swr';
 import useEtherSWR from './useEtherSWR';
+import ERC20GuildContract from 'contracts/ERC20Guild.json';
 
 interface UseVotingPowerOfProps {
   contractAddress: string;
@@ -21,5 +22,8 @@ export const useVotingPowerOf: UseVotingPowerOfHook = ({
   useEtherSWR(
     contractAddress && userAddress
       ? [contractAddress, 'votingPowerOf', userAddress]
-      : []
+      : [],
+    {
+      ABIs: new Map([[contractAddress, ERC20GuildContract.abi]]),
+    }
   );
